Next Major Update
In this update the database has been modified greatly. The database recognizes permission changes and file deletions.

Known issues with this version are that if the parent file changes, the database will not recognize it. Additionally, if the application code breaks in the middle, the database will have issues with temporary tables, that will not be dropped between runs. To fix this the user must connect to the database and  use SQLITE3 "DROP TABLE $$$$" to drop the table causing the issue. Cloudperm.py has a function called Drop that can be used to drop the temporary tables as well. From test cases, data integrity seems to be intact at this point in time.